Simplify implementation and rename test
